Revert "Personal Management Screen UI implementation"
