--- lib/json/post.json
@@ -1,20 +1,0 @@
-[
-  {
-    "id": 1,
-    "title": "",
-    "author": "",
-    "content": "",
-    "banner": "",
-    "platform": {
-      "id": 1,
-      "name": "",
-      "logo": ""
-    },
-    "createdAt":"2020-05-12T18:44:04.216Z",
-    "updatedAt":"2020-05-14T18:44:04.216Z",
-    "like": 4,
-    "dislike": 1,
-    "saved": 2,
-    "comments": [1,2]
-  }
-]
