--- lib/json/post_read.json
@@ -1,9 +1,0 @@
-{
-  "id": 1,
-  "post_id": 1,
-  "user_id": 1,
-  "createdAt":"2020-06-11T18:44:04.216Z",
-  "updatedAt":"2020-06-14T18:44:04.216Z",
-  "percentage": 85,
-  "comments": [1, 2]
-}
