--- lib/json/user.json
@@ -1,49 +1,0 @@
-[
-  {
-    "id": 1,
-    "authentication_uid": "1M6piZQhwFeum59zwiaIhXd2ZnS2",
-    "name": "凯瑟琳",
-    "photo": "",
-    "weChat": "",
-    "facebook": "",
-    "liked":[2, 3],
-    "be_liked": [3],
-    "createdAt":"2020-06-12T18:44:04.216Z",
-    "updatedAt":"2020-06-14T18:44:04.216Z",
-    "reading_time_per_day": [
-      {
-        "time": 122,
-        "date": "2020-06-12"
-      },
-      {
-        "time": 12,
-        "date": "2020-06-13"
-      },
-      {
-        "time": 444,
-        "date": "2020-06-14"
-      }
-    ],
-    "posts_read": [1],
-    "messages": [
-      {
-        "comment_id": 3,
-        "unread": true
-      },
-      {
-        "comment_id": 2,
-        "unread": false
-      }
-    ],
-    "saved_posts": [
-      {
-        "post_id": 1,
-        "saved_date": "2020-06-14"
-      },
-      {
-        "post_id": 2,
-        "saved_date": "2020-06-12"
-      }
-    ]
-  }
-]
